backend/requirements.txt
djangorestframework
markdown  
django-filter
psycopg2
djangorestframework
djangorestframework-jwt
cryptography
python-jose
django-cors-headers